+++ package.js
@@ -23,20 +23,19 @@ Package.onTest( function( api ){
 });
 
 function configure( api ){
-    api.versionsFrom( '2.9.0' );
-    api.use( 'accounts-password@2.3.3' );
-    api.use( 'blaze-html-templates@2.0.0', 'client' );
-    api.use( 'deanius:promise@3.1.3', 'client' );
+    api.versionsFrom([ '2.9.0', '3.0-rc.0' ]);
+    api.use( 'accounts-password@2.3.3 || 3.0.0-rc300.2' );
+    api.use( 'blaze-html-templates@2.0.0 || 3.0.0-alpha300.0', 'client' );
     api.use( 'ecmascript' );
     api.use( 'less@4.0.0', 'client' );
-    api.use( 'pwix:bootbox@1.5.0' );
-    api.use( 'pwix:i18n@1.5.0' );
-    api.use( 'pwix:layout@1.3.0' );
-    api.use( 'pwix:modal@1.7.0' );
-    api.use( 'pwix:options@2.1.0' );
-    api.use( 'pwix:tolert@1.4.0' );
+    api.use( 'pwix:bootbox@1.5.5' );
+    api.use( 'pwix:i18n@1.5.7' );
+    api.use( 'pwix:layout@1.3.1' );
+    api.use( 'pwix:modal@1.10.0' );
+    api.use( 'pwix:options@2.1.1' );
+    api.use( 'pwix:tolert@1.4.2' );
     api.use( 'random', 'client' );
-    api.use( 'tmeasday:check-npm-versions@1.0.2', 'server' );
+    api.use( 'tmeasday:check-npm-versions@1.0.2 || 2.0.0-beta.0', 'server' );
     api.addFiles( 'src/client/components/acMandatoryField/acMandatoryField.js', 'client' );
     api.addFiles( 'src/client/components/acMandatoryFooter/acMandatoryFooter.js', 'client' );
     api.addFiles( 'src/client/components/acMenuItems/acMenuItems.js', 'client' );